requirements.txt
asgiref==3.9.1
click==8.2.1
coverage==7.10.5
Django==5.2.5
djangorestframework==3.16.1
drf-yasg==1.21.10
gunicorn==23.0.0
h11==0.16.0
inflection==0.5.1
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0
uvicorn==0.35.0
